cast no longer necessary
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@38668 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/vector/rtree/index.c
@@ -126,7 +126,7 @@ int RTreeSearch(struct RTree *t, struct Rect *r, SearchHitCallback shcb,
 		    RTreeOverlap(r, &(s[top].sn->branch[i].rect), t)) {
 		    hitCount++;
 		    if (shcb) {	/* call the user-provided callback */
-			if (!shcb((int)s[top].sn->branch[i].child.id, cbarg)) {
+			if (!shcb(s[top].sn->branch[i].child.id, cbarg)) {
 			    /* callback wants to terminate search early */
 			    return hitCount;
 			}
